Move the partition builder spec into its own file
